requirements.txt
pytest==7.4.0
allure-pytest==2.13.2
requests==2.31.0
jsonschema==4.19.0
dataclasses-jsonschema==2.16.0
PyHamcrest==2.0.4
pytest-retry==1.3.0